+++ src/disseminate/tags/eqs.py
@@ -3,13 +3,10 @@ Tags to render equations
 """
 from copy import copy
 
-from .tag import TagFactory
 from .img import RenderedImg
 from .exceptions import TagError
-from .receivers import process_content, process_macros
 from .utils import content_to_str
 from ..formats import tex_cmd, tex_env
-from ..attributes import Attributes
 
 
 class Eq(RenderedImg):
@@ -34,10 +31,10 @@ class Eq(RenderedImg):
 
     aliases = ('term', 'termb')
     active = True
+    input_format = '.tex'
 
-    process_content = False
+    process_content = True
 
-    block_equation = None
     bold = None
     color = None
 
@@ -46,67 +43,61 @@ class Eq(RenderedImg):
 
     default_block_env = "align*"
 
+    _block_equation = False
+
     def __init__(self, name, content, attributes, context,
                  block_equation=False):
-        assert (context.is_valid('renderers') and
-                'equation' in context['renderers'])
         self.block_equation = block_equation
 
-        # Ensure the attributes argument is an Attributes type
-        attributes = Attributes(attributes)
-
-        # Get various properties from the attributes
-        env = attributes.get('env', target='.tex')
-        self.block_equation = (True if env is not None or
-                               block_equation else False)
-        self.tex_env = env if env is not None else self.default_block_env
-
-        # Prepare the tag to only have its contents processed
-        self.context = context
-        self.content = content
-        self.attributes = attributes
-
-        # Replace macros and process content
-        process_macros(tag=self)
-        process_content(tag=self, tag_factory=TagFactory)
-
-        # Take the content, and convert it to latex. Save the tex_content
-        # separately because the contents of this tag will be converted to
-        # an image path by the parent class.
-        tex_content = content_to_str(self.content, target='.tex', mathmode=True)
-        tex_content = tex_content.strip(' \t\n')
-        self.tex_content = tex_content
-        content = self.tex
-
-        # Crop equation images created by the dependency manager. This removes
-        # white space around the image so that the equation images.
-        # Note: This crop command should not cut off baselines such that
-        # equation images won't line up properly with the surrounding text.
-        # ex: H vs Hy
-        attributes['crop'] = (100, 100, 0, 0)
-
         super().__init__(name=name, content=content, attributes=attributes,
-                         context=context, render_target='.tex')
+                         context=context)
 
-    def render_content(self, content, context):
+        # Convert the content into a string formatted for latex
+        content = content_to_str(self.content, target='.tex',
+                                 mathmode=True)
+        self.content = content.strip(' \t\n')
+
+    @property
+    def block_equation(self):
+        return (self._block_equation or
+                'env' in self.attributes or
+                self.paragraph_role == 'block')
+
+    @block_equation.setter
+    def block_equation(self, value):
+        self._block_equation = value
+
+    def prepare_content(self, content=None, context=None):
         """Render the content in LaTeX into a valid .tex file."""
+        context = context or self.context
+
         # Get the renderer for the equation
         if ('renderers' not in context or
            'equation' not in context['renderers']):
             raise TagError("Missing equation renderer in the document context")
         renderer = context['renderers']['equation']
 
+        # Get the unspecified arguments
+        # FIXME: content or tex_fmt
+        content = self.tex_fmt(mathmode=False)
+        context = context if context is not None else context
+
         # Make a copy of the context and add the content as the 'body' entry.
         context_cp = copy(context)
         context_cp['body'] = content
-
-        # Get the kwargs and args to pass to the template
-        kwargs = self.template_kwargs()
+        kwargs = dict()
 
         return renderer.render(context=context_cp, target='.tex', **kwargs)
 
     def html_fmt(self, content=None, attributes=None, level=1):
-        attrs = attributes if attributes is not None else self.attributes.copy()
+        attrs = attributes or self.attributes.copy()
+
+        # Crop equation images created by the dependency manager. This removes
+        # white space around the image so that the equation images.
+        # Note: This crop command should not cut off baselines such that
+        # equation images won't line up properly with the surrounding text.
+        # ex: H vs Hy
+        attrs['crop'] = (100, 100, 0, 0)
 
         if self.block_equation:
             attrs['class'] = 'eq blockeq'
@@ -116,12 +107,14 @@ class Eq(RenderedImg):
                                         level=level)
 
     def tex_fmt(self, content=None, attributes=None, mathmode=False, level=1):
-        if content is None:
-            content = (self.tex_content if self.tex_content is not None
-                       else self.content)
+        # Retrieve unspecified arguments
+        attributes = attributes or self.attributes
+        content = content or self.content
+
+        # Determine the environment
+        env = attributes.get('env', target='.tex') or self.default_block_env
 
         # Add bold and color if specified
-        attributes = attributes if attributes is not None else self.attributes
         if 'color' in attributes:
             content = tex_cmd(cmd='textcolor',
                               attributes=attributes['color'],
@@ -135,8 +128,8 @@ class Eq(RenderedImg):
         if mathmode:
             return content
         else:
-            if self.block_equation or self.paragraph_role == 'block':
-                return tex_env(env=self.tex_env, attributes=attributes,
+            if self.block_equation:
+                return tex_env(env=env, attributes=attributes,
                                formatted_content=content, min_newlines=True)
             else:
                 return tex_cmd(cmd='ensuremath', attributes=attributes,